refactor(KERN): remove legacy design checkbox (#3618)
