+++ app/components/formElements/schemaToForm/renderZodEnum.tsx
@@ -4,7 +4,7 @@ import type { StrapiFormComponent } from "~/services/cms/models/formElements/Str
 import { sortSchemaOptionsByFormComponents } from "./sortSchemaOptionsByFormComponents";
 import KernRadioGroup from "~/components/kern/formElements/KernRadioGroup";
 import KernTile from "~/components/kern/formElements/tile/KernTile";
-import KernCheckbox from "~/components/kern/formElements/KernCheckbox";
+import KernCheckbox from "~/components/formElements/KernCheckbox";
 import KernSelect from "~/components/kern/formElements/KernSelect";
 
 type ZodEnum = z.ZodEnum<Record<string, string>>;
