+++ app/services/pdf/checkboxListToString.ts
@@ -1,4 +1,4 @@
-import { type CheckboxValue } from "~/components/formElements/Checkbox";
+import { type CheckboxValue } from "~/components/formElements/KernCheckbox";
 
 export function checkboxListToString(
   mapping: Record<string, string>,
